+++ ConstructionSite.Extensions/Images/ControllerImageExtension.cs
@@ -2,15 +2,12 @@
 using ConstructionSite.Repository.Abstract;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace ConstructionSite.Extensions.Images
 {
-  public static   class ControllerImageExtension
+    public static   class ControllerImageExtension
     {
         public async static Task<int> SaveImage(this IFormFile file, IWebHostEnvironment _env, string subFolder,Image image, IUnitOfWork _unitOfWork)
         {
