+++ ConstructionSite.Extensions/Images/ImageExtensions.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 
@@ -9,21 +10,14 @@ namespace ConstructionSite.Extensions.Images
 {
     public static class ImageExtensions
     {
-        public static bool IsImage(this IFormFile file)
-        {
-            return file.ContentType == "image/jpeg" ||
-                   file.ContentType == "image/jpg" ||
-                   file.ContentType == "image/png" ||
-                   file.ContentType == "image/x-png" ||
-                   file.ContentType == "image/gif";
-        }
+       
 
         public static async Task<string> SaveAsync(this IFormFile file, IWebHostEnvironment _env, string subFolder)
         {
 
+           
 
-          
-                string fileName = Imager.GetImageSubFolder(subFolder, file.FileName);
+            string fileName = Imager.GetImageSubFolder(subFolder, file.FileName);
 
                 string path = Path.Combine(_env.WebRootPath, "images", fileName);
 
@@ -35,5 +29,21 @@ namespace ConstructionSite.Extensions.Images
                 return fileName;
            
         }
+        //public static async Task<string> SaveAsyncWith(this List<IFormFile> files)
+        //{
+        //    //string fileName = Imager.GetImageSubFolder(subFolder, file.FileName);
+         
+        //    string path = Path.Combine(_env.WebRootPath, "images", fileName);
+        //    foreach (var formFile in files)
+        //    {
+        //        if (formFile.Length > 0)
+        //        {
+        //            using (var stream = new FileStream(path, FileMode.Create))
+        //            {
+        //                await formFile.CopyToAsync(stream);
+        //            }
+        //        }
+        //    }
+        //}
     }
 }
